refactor: 관심사 매핑하는 로직과 DTO rename
