+++ Popcorn-iOS/Source/Domain/UseCases/SignUpUseCase/SignUpUseCase.swift
@@ -46,20 +46,7 @@ final class SignUpUseCase: SignUpUseCaseProtocol {
 
     // MARK: - Private func
     private func convertInterestToEnglish(_ interest: String) -> String {
-        let mapping: [String: String] = [
-            "패션": "FASHION",
-            "뷰티": "BEAUTY",
-            "음식": "FOOD",
-            "캐릭터": "CHARACTER",
-            "드라마/영화": "MOVIES",
-            "라이프 스타일": "LIFESTYLE",
-            "예술": "ART",
-            "IT": "IT",
-            "스포츠": "SPORTS",
-            "셀럽": "CELEBRITY",
-            "반려동물": "PETS"
-        ]
-        return mapping[interest] ?? interest
+        return InterestCategory(rawValue: interest)?.serverValue ?? interest
     }
 }
 
@@ -131,8 +118,8 @@ extension SignUpUseCase {
 
     func saveSignUpData(name: String, id: String, password: String, email: String) -> Bool {
         let data = SignUpRequestDTO(
-            firstSignupDto: FirstSignupDto(name: name, username: id, password: password, email: email),
-            secondSignupDto: nil
+            firstSignupDTO: FirstSignupDTO(name: name, username: id, password: password, email: email),
+            secondSignupDTO: nil
         )
         return signUpRepository.saveSignUpData(signUpData: data)
     }
@@ -188,8 +175,8 @@ extension SignUpUseCase {
         }
         let convertedInterests = interests.map { convertInterestToEnglish($0) }
         let updateSignUpData = SignUpRequestDTO(
-            firstSignupDto: firstSignUpData.firstSignupDto,
-            secondSignupDto: SecondSignupDto(nickname: nickName, profileId: profileId, interests: convertedInterests)
+            firstSignupDTO: firstSignUpData.firstSignupDTO,
+            secondSignupDTO: SecondSignupDTO(nickname: nickName, profileId: profileId, interests: convertedInterests)
         )
 
         signUpRepository.fetchSignUpResult(signupData: updateSignUpData) { result in
